Separate mappers and repositories by packages
