+++ app/src/main/java/com/simply/birthdayapp/data/mappers/ShopMapper.kt
@@ -1,4 +1,4 @@
-package com.simply.birthdayapp.data.shops
+package com.simply.birthdayapp.data.mappers
 
 import android.util.Base64
 import com.simply.birthdayapp.ShopsQuery
